Move tests to where they belong
